refactored absolute db notifications
